requirements.txt
setuptools
django-suit
django-extensions
django-localflavor-br
pytz
django-money
sendgrid
django-ckeditor
django-admin-tools
Pillow
sudo apt-get install python-pip python-dev libmysqlclient-dev
MySQL-python
djangorestframework
coreapi
simplejson
Markdown	
django-nested-inline
django-multiselectfield
django-grappelli
djangorestframework-filters
coreapi-cli
django-phonenumber-field
djangorestframework-word-filter
django-filter
django-guardian
django-crispy-forms
pymongo
django-image-cropping
mongoadmin
mongodbforms
django-mongo-sessions
tox
psycopg2
